requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
imbalanced-learn
sentence-transformers
vaderSentiment
transformers
datasets
accelerate
pyyaml
mlflow
python-dotenv
dvc


huggingface_hub